use transaction for db unit test (only teams)

+++ backend/util/random.go
@@ -9,8 +9,8 @@ import (
 
 const alphabet = "abcdefghijklmnopqrstuvwxyz"
 
-func RandomInt(min, max int64) int64 {
-	return min + rand.Int63n(max-min+1)
+func RandomInt(min, max int64) int32 {
+	return int32(min + rand.Int63n(max-min+1))
 }
 
 func RandomString(n int) string {
